pyproject.toml
[project]
name = "g-daily"
version = "1.0.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "keyboard>=0.13.5",
    "pyinstaller>=6.19.0",
    "pyqtdarktheme>=2.1.0",
    "pyside6-essentials>=6.10.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
